medium sized refactor
